+++ pages/status.tsx
@@ -5,8 +5,8 @@ import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { useFormik } from 'formik'
 import * as Yup from 'yup'
-import { CheckStatusRequest } from '../lib/StatusTypes'
-import { useCheckStatus } from '../lib/CheckStatusHook'
+import { CheckStatusRequest } from '../lib/types'
+import { useCheckStatus } from '../lib/useCheckStatus'
 import Layout from '../components/Layout'
 import InputField from '../components/InputField'
 import ActionButton from '../components/ActionButton'
